requirements.txt
gunicorn
cython
flask
pandas
xgboost
tensorflow-cpu
nltk
scikit-learn
